backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
motor==3.6.0
redis==5.2.0
pyjwt[crypto]==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic[email]==2.9.2
slowapi==0.1.9
limits==3.13.0
httpx==0.27.2
python-dateutil==2.9.0
